pyproject.toml
[project]
name = "lab4_bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = [
    "keras>=3.0.5",
    "python-telegram-bot>=21.0.1",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "GPL-3.0"}

[tool.pdm.scripts]
start.cmd = "python -m lab4_bot"
start.env_file = ".env"
